test: get the two process instance archiver jobs to use same tests
they should behave in the same manner (from a functional POV) so we
might as well test them in the same way.
